Removed floating-point computations, and simplified calculations significantly.

+++ GV/Drivers/ISpeedSensor.h
@@ -13,7 +13,7 @@
 class ISpeedSensor : public IDevice {
 public:
 
-	virtual float getSpeed() = 0;
+	virtual int getSpeed() = 0;
 };
 
 #endif /* ISPEEDSENSOR_H_ */
